requirements.txt
absl-py==2.1.0
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deepface==0.0.93
Django==5.1.4
django-allauth==65.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-auth==2.37.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
kombu==5.4.2
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
mysqlclient==2.2.7
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.48
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
scikit-image==0.25.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2025.1.10
timedelta==2020.12.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.17.2
